+++ src/HtmlPageCrawler.php
@@ -1097,4 +1097,13 @@ class HtmlPageCrawler extends Crawler
         $parent = $this->getNode(0)->parentNode;
         return ($parent == null || $parent->tagName == self::FRAGMENT_ROOT_TAGNAME);
     }
+
+    public function __get($name) {
+        switch($name) {
+            case 'count':
+            case 'length':
+                return count($this);
+        }
+        throw new \Exception('No such property ' . $name);
+    }
 }